QuotesApp/backend/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-OAuthlib==0.9.6
Flask-Mail==0.9.1
mysql-connector-python==8.0.34
python-dotenv==1.0.0